service/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-seed==0.3.1
djangorestframework==3.15.2
Faker==26.0.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
toposort==1.10
python-dotenv==1.0.1